+++ src/GreenFunc.jl
@@ -2,6 +2,13 @@ module GreenFunc
 using StaticArrays, Lehmann, CompositeGrids#, BZMeshes
 # Write your package code here.
 
+include("green/meshgrids/MeshGrids.jl")
+using .MeshGrids
+export MeshGrids
+export locate, volume
+export FERMI, BOSE, UNKNOWN
+export MeshProduct
+export DLRFreq
 
 include("green/Green.jl")
 export TimeDomain, ImTime, ReTime, ImFreq, ReFreq, DLRFreq
@@ -13,14 +20,11 @@ export GreenDLR
 include("green/GreenSym.jl")
 export GreenSym2DLR, dynamic, instant
 
-include("green/meshgrids/MeshProduct.jl")
-export MeshProduct
-export locate, volume
+# include("green/meshgrids/MeshProduct.jl")
+# export MeshProduct
+# export locate, volume
 
 include("green/GreenNew.jl")
 export GreenNew
-
-include("green/meshgrids/MeshGrids.jl")
-export MeshGrids
 
 end
